fix: solve final cyclical dep
